Deploy ioncube only if the php_deployment var is set

+++ deploy.sh
@@ -167,39 +167,41 @@ fi
 
 sysLogger "TEXT" "\n###################################\n#        Plesk PHP Ioncube        #\n###################################\n";
 TMP_IONCUBE_LOADER_INI_PATH=/opt/plesk/php/7.0/etc/php.d/00-ioncube-loader.ini
-# PHP 7.0 Ioncube Deployment
-if [[ $PHP70_IONCUBE == 1 && -f /opt/plesk/php/7.0/etc/php.ini ]]; then
-	if [[ $LINUX_MACHINE_TYPE == "i686" || $LINUX_MACHINE_TYPE == "x86" ]]; then
-		# Linux x86 Systems
-		if [[ ! -d /opt/plesk/php/7.0/lib ]]; then sysLogger "ERROR" "The Ioncube Installation failed. The folder /opt/plesk/php/7.0/lib/ does not exist."; fi
-		cp $FILES_PATH/ioncube_loaders_lin_x86-32/ioncube_loader_lin_7.0.so /opt/plesk/php/7.0/lib/php/modules/ioncube_loader.so | tee -a $LOG_DEPLOYMENT;
-		printf "zend_extension=/opt/plesk/php/7.0/lib/php/modules/ioncube_loader.so" > $TMP_IONCUBE_LOADER_INI_PATH
-		chmod 755 /opt/plesk/php/7.0/lib/php/modules/ioncube_loader.so
+if [[ $PHP_DEPLOYMENT == 1 ]]; then
+	# PHP 7.0 Ioncube Deployment
+	if [[ $PHP70_IONCUBE == 1 && -f /opt/plesk/php/7.0/etc/php.ini ]]; then
+		if [[ $LINUX_MACHINE_TYPE == "i686" || $LINUX_MACHINE_TYPE == "x86" ]]; then
+			# Linux x86 Systems
+			if [[ ! -d /opt/plesk/php/7.0/lib ]]; then sysLogger "ERROR" "The Ioncube Installation failed. The folder /opt/plesk/php/7.0/lib/ does not exist."; fi
+			cp $FILES_PATH/ioncube_loaders_lin_x86-32/ioncube_loader_lin_7.0.so /opt/plesk/php/7.0/lib/php/modules/ioncube_loader.so | tee -a $LOG_DEPLOYMENT;
+			printf "zend_extension=/opt/plesk/php/7.0/lib/php/modules/ioncube_loader.so" > $TMP_IONCUBE_LOADER_INI_PATH
+			chmod 755 /opt/plesk/php/7.0/lib/php/modules/ioncube_loader.so
+			plesk bin php_handler --reread | tee -a $LOG_DEPLOYMENT;
+			sysLogger "DONE" "The Ioncube PHP 7.0 Loader is successfully deployed.";
+		elif [[ $LINUX_MACHINE_TYPE == "x86_64" ]]; then
+			# Linux x86_64 Systems
+			if [[ ! -d /opt/plesk/php/7.0/lib64 ]]; then sysLogger "ERROR" "The Ioncube Installation failed. The folder /opt/plesk/php/7.0/lib64/ does not exist."; fi
+			cp $FILES_PATH/ioncube_loaders_lin_x86-64/ioncube_loader_lin_7.0.so /opt/plesk/php/7.0/lib64/php/modules/ioncube_loader.so | tee -a $LOG_DEPLOYMENT;
+			printf "zend_extension=/opt/plesk/php/7.0/lib64/php/modules/ioncube_loader.so" > $TMP_IONCUBE_LOADER_INI_PATH
+			chmod 755 /opt/plesk/php/7.0/lib64/php/modules/ioncube_loader.so
+			plesk bin php_handler --reread | tee -a $LOG_DEPLOYMENT;
+			sysLogger "DONE" "The Installation of the PHP 7.0 Ioncube Loader was successful (please check if there are any possible errors above).";
+		else
+			# Linux System Type not found
+			sysLogger "ERROR" "The Installation of the PHP 7.0 Ioncube Loader failed. The Plesk Deployer wasn't able to determine your Linux Machine Type (x86 or x86_64).";
+		fi
+	elif [[ $PHP70_IONCUBE == 1 && ! -f /opt/plesk/php/7.0/etc/php.ini ]]; then
+		# PHP 7.0 not installed on this system
+		sysLogger "WARNING" "PHP7.0 is not installed on this system, therefore a Ioncube Deployment for PHP7 is not possible.";
+	elif [[ $PHP70_IONCUBE == 0 && -f $TMP_IONCUBE_LOADER_INI_PATH ]]; then
+		# Uninstall the PHP 7.0 Ioncube Loader
+		rm -f $TMP_IONCUBE_LOADER_INI_PATH
 		plesk bin php_handler --reread | tee -a $LOG_DEPLOYMENT;
-		sysLogger "DONE" "The Ioncube PHP 7.0 Loader is successfully deployed.";
-	elif [[ $LINUX_MACHINE_TYPE == "x86_64" ]]; then
-		# Linux x86_64 Systems
-		if [[ ! -d /opt/plesk/php/7.0/lib64 ]]; then sysLogger "ERROR" "The Ioncube Installation failed. The folder /opt/plesk/php/7.0/lib64/ does not exist."; fi
-		cp $FILES_PATH/ioncube_loaders_lin_x86-64/ioncube_loader_lin_7.0.so /opt/plesk/php/7.0/lib64/php/modules/ioncube_loader.so | tee -a $LOG_DEPLOYMENT;
-		printf "zend_extension=/opt/plesk/php/7.0/lib64/php/modules/ioncube_loader.so" > $TMP_IONCUBE_LOADER_INI_PATH
-		chmod 755 /opt/plesk/php/7.0/lib64/php/modules/ioncube_loader.so
-		plesk bin php_handler --reread | tee -a $LOG_DEPLOYMENT;
-		sysLogger "DONE" "The Installation of the PHP 7.0 Ioncube Loader was successful (please check if there are any possible errors above).";
+		sysLogger "DONE" "The Uninstallation of the PHP 7.0 Ioncube Loader was successful.";
 	else
-		# Linux System Type not found
-		sysLogger "ERROR" "The Installation of the PHP 7.0 Ioncube Loader failed. The Plesk Deployer wasn't able to determine your Linux Machine Type (x86 or x86_64).";
+		# No PHP 7.0 Deploymet specified
+		sysLogger "INFO" "No Deployment for the PHP 7.0 Ioncube Loader specified (skip).";
 	fi
-elif [[ $PHP70_IONCUBE == 1 && ! -f /opt/plesk/php/7.0/etc/php.ini ]]; then
-	# PHP 7.0 not installed on this system
-	sysLogger "WARNING" "PHP7.0 is not installed on this system, therefore a Ioncube Deployment for PHP7 is not possible.";
-elif [[ $PHP70_IONCUBE == 0 && -f $TMP_IONCUBE_LOADER_INI_PATH ]]; then
-	# Uninstall the PHP 7.0 Ioncube Loader
-	rm -f $TMP_IONCUBE_LOADER_INI_PATH
-	plesk bin php_handler --reread | tee -a $LOG_DEPLOYMENT;
-	sysLogger "DONE" "The Uninstallation of the PHP 7.0 Ioncube Loader was successful.";
-else
-	# No PHP 7.0 Deploymet specified
-	sysLogger "INFO" "No Deployment for the PHP 7.0 Ioncube Loader specified (skip).";
 fi
 
 sysLogger "TEXT" "\n###################################\n# Export Default / Custom Scripts #\n###################################\n";
